requirements.txt
python-dotenv==1.0.0
openai==0.27.8
langchain==0.0.271
PyPDF2==3.0.1
streamlit==1.26.0
fastapi==0.85.1
uvicorn==0.22.0
chromadb==0.3.29
tiktoken==0.4.0
pypdf==3.12.1